Moved several files around for better organization.
