+++ src/App.js
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react'
 import Navbar from './Components/Navbar'
 
 import './App.css'
-import SignUp from './Components/SignUp'
-import Login from './Components/Login'
-import Home from './Components/Home'
+import SignUp from './Components/Home/SignUp'
+import Login from './Components/Home/Login'
+import Home from './Components/Home/Home'
 import Booking from './Components/Booking'
 import { BrowserRouter, Route, Routes, useParams } from 'react-router-dom'
 import { getAuth, onAuthStateChanged } from "firebase/auth";
